requirements.txt
anthropic>=0.40.0
fastapi>=0.110.0
uvicorn>=0.29.0
playwright>=1.43.0
chromadb>=0.5.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
python-multipart>=0.0.9
aiofiles>=23.2.1
apscheduler>=3.10.4
requests>=2.31.0
pdfplumber>=0.10.0
pypdf>=4.0.0
docx2txt>=0.8
beautifulsoup4>=4.12.0
httpx>=0.27.0
slowapi>=0.1.9
